Update openpgp-key.ts
Typo!

'skiped' should be 'skipped'

This is done because otherwise, the form vowel + consonant + e causes the vowel to take the long form (eg the words Skype, hope, or dine). Consider the word hop. To prevent it's past simple from being 'hoped' (which is pronounced with the long o and means something else), it becomes hopped.

Interesting to me how these things kind of have rules, but also kind of don't!